Refactored the app's design and structure
Added the base files and structure to build out the new web app's frontend
