+++ src/app/(app shell)/schedule/page.tsx
@@ -1,0 +1,13 @@
+"use client";
+
+import React from "react";
+
+export default function SchedulePage() {
+	return (
+		<main className="min-h-[60vh]">
+			<h1 className="sr-only">Schedule</h1>
+			<div className="p-6 text-sm text-gray-500">
+			</div>
+		</main>
+	);
+}
