--- src/app/components/icons/Cube.tsx
@@ -1,20 +1,0 @@
-import React from 'react';
-
-const Cube = () => (
-  <svg width="12" height="12" viewBox="0 0 12 12" fill="none" xmlns="http://www.w3.org/2000/svg">
-    <g clipPath="url(#clip0_1_4136)">
-      <path d="M10.5 3L6 0.5L1.5 3L6 5.5L10.5 3Z" stroke="#6F00FF" strokeWidth="0.75" strokeLinecap="round" strokeLinejoin="round"/>
-      <path d="M1.5 8.5L6 11L10.5 8.5" stroke="#6F00FF" strokeWidth="0.75" strokeLinecap="round" strokeLinejoin="round"/>
-      <path d="M1.5 3V8.5" stroke="#6F00FF" strokeWidth="0.75" strokeLinecap="round" strokeLinejoin="round"/>
-      <path d="M10.5 3V8.5" stroke="#6F00FF" strokeWidth="0.75" strokeLinecap="round" strokeLinejoin="round"/>
-      <path d="M6 5.5V11" stroke="#6F00FF" strokeWidth="0.75" strokeLinecap="round" strokeLinejoin="round"/>
-    </g>
-    <defs>
-      <clipPath id="clip0_1_4136">
-        <rect width="12" height="12" fill="white"/>
-      </clipPath>
-    </defs>
-  </svg>
-);
-
-export default Cube;
