+++ src/app/layout.tsx
@@ -1,40 +1,29 @@
-import type { Metadata } from "next";
-import { Geist, Geist_Mono } from "next/font/google";
-import { Analytics } from "@vercel/analytics/react";
-import { SpeedInsights } from "@vercel/speed-insights/next";
-import { inter, playfair, robotoMono, notoSerif, fontVariables } from "./config/fonts";
-import Navbar from "./components/Navbar";
-import "./globals.css";
-
-const geistSans = Geist({
-  variable: "--font-geist-sans",
-  subsets: ["latin"],
-});
-
-const geistMono = Geist_Mono({
-  variable: "--font-geist-mono",
-  subsets: ["latin"],
-});
+import type { Metadata, Viewport } from "next";
+import { gilroy } from "@/config/fonts";
+import "@/app/globals.css";
+import Sidebar from "@/app/(app shell)/sidebar/Sidebar";
 
 export const metadata: Metadata = {
-  title: "SABC - Rowing Club Management",
-  description: "St Antony's College, Oxford Boat Club - Manage outings, member assignments, and availability",
+  title: "SABC — Outings",
+  description: "Sign up and confirm availability for water outings, erg, gym, and tank sessions.",
 };
 
-export default function RootLayout({
-  children,
-}: Readonly<{
-  children: React.ReactNode;
-}>) {
+export const viewport: Viewport = {
+  width: "device-width",
+  initialScale: 1,
+  maximumScale: 1,
+};
+
+export default function RootLayout({ children }: { children: React.ReactNode }) {
   return (
-    <html lang="en" className={`${inter.variable} ${notoSerif.variable} ${playfair.variable} ${robotoMono.variable}`}>
-      <body
-        className="font-inter antialiased min-h-screen px-4 sm:px-8 md:px-12 lg:px-16"
-      >
-        <Navbar />
-        {children}
-        <Analytics />
-        <SpeedInsights />
+    <html lang="en" className={gilroy.variable}>
+      <body className="min-h-screen bg-bg text-foreground antialiased">
+        <div className="flex min-h-screen">
+          <Sidebar />
+          <main className="flex-1 min-h-screen max-h-screen overflow-y-auto bg-surface p-6">
+            {children}
+          </main>
+        </div>
       </body>
     </html>
   );
