+++ src/components/icons/AvailableIcon.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+export default function AvailableIcon() {
+  return (
+    <svg width="20" height="22" viewBox="0 0 13 14" fill="none" xmlns="http://www.w3.org/2000/svg">
+      <path d="M2.56944 7.51389L4.625 9.56944L9.76389 4.43055" stroke="#00C53E" strokeWidth="1.5" strokeLinecap="round" strokeLinejoin="round"/>
+    </svg>
+  );
+}
