+++ src/components/icons/CheckCircle.tsx
@@ -1,0 +1,17 @@
+import React from 'react';
+
+const CheckCircle = () => (
+  <svg width="12" height="12" viewBox="0 0 12 12" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <g clipPath="url(#clip0_1_4133)">
+      <path d="M3.5 6.25L5 7.75L8.5 4.25" stroke="#6F00FF" strokeWidth="0.75" strokeLinecap="round" strokeLinejoin="round"/>
+      <path d="M6 11C8.76142 11 11 8.76142 11 6C11 3.23858 8.76142 1 6 1C3.23858 1 1 3.23858 1 6C1 8.76142 3.23858 11 6 11Z" stroke="#6F00FF" strokeWidth="0.75" strokeLinecap="round" strokeLinejoin="round"/>
+    </g>
+    <defs>
+      <clipPath id="clip0_1_4133">
+        <rect width="12" height="12" fill="white"/>
+      </clipPath>
+    </defs>
+  </svg>
+);
+
+export default CheckCircle;
