+++ src/components/icons/ColorWheel.tsx
@@ -1,0 +1,25 @@
+import React from 'react';
+
+const ColorWheel = () => (
+  <svg width="12" height="12" viewBox="0 0 12 12" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <g clipPath="url(#clip0_1_4139)">
+      <path d="M10 5C10 7.76142 7.76142 10 5 10C2.23858 10 0 7.76142 0 5C0 2.23858 2.23858 0 5 0C7.76142 0 10 2.23858 10 5Z" stroke="#6F00FF" strokeWidth="0.75" strokeLinecap="round" strokeLinejoin="round"/>
+      <path d="M5 2.5C3.61929 2.5 2.5 3.61929 2.5 5C2.5 6.38071 3.61929 7.5 5 7.5C6.38071 7.5 7.5 6.38071 7.5 5C7.5 3.61929 6.38071 2.5 5 2.5Z" stroke="#6F00FF" strokeWidth="0.75" strokeLinecap="round" strokeLinejoin="round"/>
+      <path d="M5 0V2.5" stroke="#6F00FF" strokeWidth="0.75" strokeLinecap="round" strokeLinejoin="round"/>
+      <path d="M5 7.5V10" stroke="#6F00FF" strokeWidth="0.75" strokeLinecap="round" strokeLinejoin="round"/>
+      <path d="M2.5 5H0" stroke="#6F00FF" strokeWidth="0.75" strokeLinecap="round" strokeLinejoin="round"/>
+      <path d="M10 5H7.5" stroke="#6F00FF" strokeWidth="0.75" strokeLinecap="round" strokeLinejoin="round"/>
+      <path d="M3.5625 3.5625L1.75 1.75" stroke="#6F00FF" strokeWidth="0.75" strokeLinecap="round" strokeLinejoin="round"/>
+      <path d="M8.25 8.25L6.4375 6.4375" stroke="#6F00FF" strokeWidth="0.75" strokeLinecap="round" strokeLinejoin="round"/>
+      <path d="M1.75 8.25L3.5625 6.4375" stroke="#6F00FF" strokeWidth="0.75" strokeLinecap="round" strokeLinejoin="round"/>
+      <path d="M6.4375 3.5625L8.25 1.75" stroke="#6F00FF" strokeWidth="0.75" strokeLinecap="round" strokeLinejoin="round"/>
+    </g>
+    <defs>
+      <clipPath id="clip0_1_4139">
+        <rect width="12" height="12" fill="white"/>
+      </clipPath>
+    </defs>
+  </svg>
+);
+
+export default ColorWheel;
