+++ src/components/icons/MaybeIcon.tsx
@@ -1,0 +1,10 @@
+import React from "react";
+
+export default function MaybeIcon() {
+  return (
+    <svg width="20" height="22" viewBox="0 0 13 14" fill="none" xmlns="http://www.w3.org/2000/svg">
+      <path d="M6.51333 10.5565C6.13861 10.5565 5.83833 10.2563 5.83833 9.88152C5.83833 9.50681 6.13861 9.20652 6.51333 9.20652C6.88804 9.20652 7.18833 9.50681 7.18833 9.88152C7.18833 10.2563 6.88804 10.5565 6.51333 10.5565Z" fill="#6F00FF"/>
+      <path d="M6.51333 7.85652V4.4873" stroke="#6F00FF" strokeWidth="1.5" strokeLinecap="round"/>
+    </svg>
+  );
+}
