+++ src/components/icons/UnavailableIcon.tsx
@@ -1,0 +1,10 @@
+import React from "react";
+
+export default function UnavailableIcon() {
+  return (
+    <svg width="20" height="22" viewBox="0 0 13 14" fill="none" xmlns="http://www.w3.org/2000/svg">
+      <path d="M9.76389 4.43056L2.56944 11.625" stroke="#8B4513" strokeWidth="1.5" strokeLinecap="round"/>
+      <path d="M2.56944 4.43056L9.76389 11.625" stroke="#8B4513" strokeWidth="1.5" strokeLinecap="round"/>
+    </svg>
+  );
+}
