+++ src/lib/classnames.ts
@@ -1,0 +1,11 @@
+// src/lib/classnames.ts
+import { clsx, type ClassValue } from "clsx";
+import { twMerge } from "tailwind-merge";
+
+/**
+ * Merges Tailwind classes conditionally.
+ * Example: cn("p-4", condition && "bg-red-500")
+ */
+export function cn(...inputs: ClassValue[]) {
+  return twMerge(clsx(inputs));
+}
